+++ src/main/java/io/odilon/virtualFileSystem/raid6/RAIDSixDriveSync.java
@@ -233,7 +233,7 @@ public class RAIDSixDriveSync implements Runnable {
 
 			for (ServerBucket bucket : getDriver().getVirtualFileSystemService().listAllBuckets()) {
 
-				Integer pageSize = Integer.valueOf(ServerConstant.DEFAULT_COMMANDS_PAGE_SIZE);
+				Long pageSize = Long.valueOf(ServerConstant.DEFAULT_COMMANDS_PAGE_SIZE);
 				Long offset = Long.valueOf(0);
 				String agentId = null;
 
@@ -241,7 +241,8 @@ public class RAIDSixDriveSync implements Runnable {
 
 				while (!done) {
 
-					DataList<Item<ObjectMetadata>> data = getDriver().getVirtualFileSystemService().listObjects(bucket.getName(), Optional.of(offset), Optional.ofNullable(pageSize), Optional.empty(), Optional.ofNullable(agentId));
+					DataList<Item<ObjectMetadata>> data = getDriver().getVirtualFileSystemService().listObjects(bucket.getName(), 
+							Optional.of(offset), Optional.ofNullable(pageSize), Optional.empty(), Optional.ofNullable(agentId));
 
 					if (agentId == null)
 						agentId = data.getAgentId();
